requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
absl-py=2.1.0=pypi_0
aiohttp=3.9.5=pypi_0
aiosignal=1.3.1=pypi_0
async-timeout=4.0.3=pypi_0
attrs=23.2.0=pypi_0
blas=1.0=mkl
c-ares=1.19.1=h5eee18b_0
ca-certificates=2024.3.11=h06a4308_0
cachetools=5.3.3=pypi_0
certifi=2024.6.2=pypi_0
charset-normalizer=3.3.2=pypi_0
cudatoolkit=11.3.1=h2bc3f7f_2
einops=0.8.0=pypi_0
fastmri=0.1.1=pypi_0
filelock=3.15.1=pypi_0
frozenlist=1.4.1=pypi_0
fsspec=2024.6.0=pypi_0
future=1.0.0=pypi_0
google-auth=2.30.0=pypi_0
google-auth-oauthlib=1.0.0=pypi_0
grpcio=1.64.1=pypi_0
h5py=3.11.0=py38hbe37b52_0
hdf5=1.12.1=h2b7332f_3
huggingface-hub=0.23.3=pypi_0
idna=3.7=pypi_0
imageio=2.34.1=pypi_0
importlib-metadata=7.1.0=pypi_0
intel-openmp=2023.1.0=hdb19cb5_46306
jinja2=3.1.4=pypi_0
krb5=1.20.1=h143b758_1
lazy-loader=0.4=pypi_0
ld_impl_linux-64=2.38=h1181459_1
libcurl=8.7.1=h251f7ec_0
libedit=3.1.20230828=h5eee18b_0
libev=4.33=h7f8727e_1
libffi=3.4.4=h6a678d5_1
libgcc-ng=11.2.0=h1234567_1
libgfortran-ng=11.2.0=h00389a5_1
libgfortran5=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libnghttp2=1.57.0=h2d74bed_0
libssh2=1.11.0=h251f7ec_0
libstdcxx-ng=11.2.0=h1234567_1
libuv=1.44.2=h5eee18b_0
lightning-utilities=0.11.2=pypi_0
markdown=3.6=pypi_0
markupsafe=2.1.5=pypi_0
mkl=2023.1.0=h213fc3f_46344
mkl-service=2.4.0=py38h5eee18b_1
mkl_fft=1.3.8=py38h5eee18b_0
mkl_random=1.2.4=py38hdb19cb5_0
mpmath=1.3.0=pypi_0
multidict=6.0.5=pypi_0
ncurses=6.4=h6a678d5_0
networkx=3.1=pypi_0
numpy=1.24.3=py38hf6e8229_1
numpy-base=1.24.3=py38h060ed82_1
nvidia-cublas-cu12=12.1.3.1=pypi_0
nvidia-cuda-cupti-cu12=12.1.105=pypi_0
nvidia-cuda-nvrtc-cu12=12.1.105=pypi_0
nvidia-cuda-runtime-cu12=12.1.105=pypi_0
nvidia-cudnn-cu12=8.9.2.26=pypi_0
nvidia-cufft-cu12=11.0.2.54=pypi_0
nvidia-curand-cu12=10.3.2.106=pypi_0
nvidia-cusolver-cu12=11.4.5.107=pypi_0
nvidia-cusparse-cu12=12.1.0.106=pypi_0
nvidia-nccl-cu12=2.20.5=pypi_0
nvidia-nvjitlink-cu12=12.5.40=pypi_0
nvidia-nvtx-cu12=12.1.105=pypi_0
oauthlib=3.2.2=pypi_0
openssl=3.0.13=h7f8727e_2
packaging=24.1=pypi_0
pillow=10.3.0=pypi_0
pip=24.0=py38h06a4308_0
protobuf=5.27.1=pypi_0
pyasn1=0.6.0=pypi_0
pyasn1-modules=0.4.0=pypi_0
python=3.8.19=h955ad1f_0
pytorch-lightning=1.9.4=pypi_0
pytorch-mutex=1.0=cuda
pywavelets=1.4.1=pypi_0
pyyaml=6.0.1=pypi_0
readline=8.2=h5eee18b_0
requests=2.32.3=pypi_0
requests-oauthlib=2.0.0=pypi_0
rsa=4.9=pypi_0
runstats=2.0.0=pypi_0
safetensors=0.4.3=pypi_0
scikit-image=0.21.0=pypi_0
scipy=1.10.1=pypi_0
setuptools=69.5.1=py38h06a4308_0
sqlite=3.45.3=h5eee18b_0
sympy=1.12.1=pypi_0
tbb=2021.8.0=hdb19cb5_0
tensorboard=2.14.0=pypi_0
tensorboard-data-server=0.7.2=pypi_0
tensorboardx=2.6.2.2=pypi_0
tifffile=2023.7.10=pypi_0
timm=1.0.3=pypi_0
tk=8.6.14=h39e8969_0
torch=2.3.1=pypi_0
torchmetrics=1.4.0.post0=pypi_0
torchvision=0.18.1=pypi_0
tqdm=4.66.4=pypi_0
triton=2.3.1=pypi_0
typing_extensions=4.11.0=py38h06a4308_0
urllib3=2.2.1=pypi_0
werkzeug=3.0.3=pypi_0
wheel=0.43.0=py38h06a4308_0
xz=5.4.6=h5eee18b_1
yarl=1.9.4=pypi_0
zipp=3.19.2=pypi_0
zlib=1.2.13=h5eee18b_1
